Function_B.java
This file contains the code for matrix operations. The code can print a submenu for the user and asks them which operations does they want to make. After the selection, the code asks the user for their matrixes and their values. According to the matrixes, transpose and matrix multiplications for the matrix can be made.